requirements.txt
awscli==1.11.127
beautifulsoup4==4.5.3
boto3==1.4.4
botocore==1.5.90
certifi==2017.4.17
chardet==3.0.4
colorama==0.3.7
contextlib2==0.5.5
Django==1.10.5
django-debug-toolbar==1.8
docutils==0.13.1
futures==3.0.5
geopy==1.11.0
gunicorn==19.7.1
idna==2.5
jmespath==0.9.1
MySQL-python==1.2.5
mysqlclient==1.3.10
olefile==0.44
Pillow==4.0.0
pyasn1==0.3.1
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
raven==6.1.0
requests==2.18.1
requests-file==1.4.2
rsa==3.4.2
s3transfer==0.1.10
six==1.10.0
sparkpost==1.3.5
sqlparse==0.2.3
stripe==1.59.0
tldextract==2.1.0
urllib3==1.21.1
virtualenv==15.1.0
